+++ LastResort/Enemy_MetalCraw.cpp
@@ -53,10 +53,10 @@ void Enemy_MetalCraw::Move()
 		numparticles += 1;
 		if (numparticles <= 1)
 		{
-			App->particles->AddParticle(App->particles->MiddleBossShot, { position.x - 10, position.y - 10 }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
-			App->particles->AddParticle(App->particles->MiddleBossShot, { position.x + 50, position.y + 10 }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
-			App->particles->AddParticle(App->particles->MiddleBossShot, { position.x - 10, position.y + 50 }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
-			App->particles->AddParticle(App->particles->MiddleBossShot, { position.x + 50, position.y + 10 }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
+			App->particles->AddParticle(App->particles->MiddleBossShot, { (float)(position.x - 10), (float)(position.y - 10) }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
+			App->particles->AddParticle(App->particles->MiddleBossShot, { (float)(position.x + 50), (float)(position.y + 10) }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
+			App->particles->AddParticle(App->particles->MiddleBossShot, { (float)(position.x - 10), (float)(position.y + 50) }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
+			App->particles->AddParticle(App->particles->MiddleBossShot, { (float)(position.x + 50), (float)(position.y + 10) }, { 0 ,0 }, App->enemies->nml_sprites, COLLIDER_ENEMY_LIGHT);
 		}
 		App->particles->VectorMiddleBossShots.x = (position.x - App->player1->position.x) / (sqrt  (pow(position.x - App->player1->position.x, 2) + pow(position.y - App->player1->position.y, 2) ));
 		App->particles->VectorMiddleBossShots.y = (position.y - App->player1->position.y) / (sqrt(pow(position.x - App->player1->position.x, 2) + pow(position.y - App->player1->position.y, 2)));
