+++ LastResort/Enemy_Outdoor_turret.cpp
@@ -32,7 +32,7 @@ void Enemy_Outdoor_turret::Move()
 	if (SDL_GetTicks() - StartTime > 3000 && position.x>0 && position.y>0)
 	{
 		/*App->enemies->AddEnemy(OUTDOOR_LASER, position_x, position_y);*/
-		App->particles->AddParticle(App->particles->LaserEnemyShot, { position.x + 7,position.y - 10 }, { 3 , 3 }, App->particles->laserTx, COLLIDER_ENEMY_SHOT_INDESTRUCTIBLE, 0, PARTICLE_LASER);
+		App->particles->AddParticle(App->particles->LaserEnemyShot, { (float)(position.x + 7), (float)(position.y - 10) }, { 3 , 3 }, App->particles->laserTx, COLLIDER_ENEMY_SHOT_INDESTRUCTIBLE, 0, PARTICLE_LASER);
 		StartTime = SDL_GetTicks();
 	}
 }
